Add logging fixture to test crate
