+++ tests/connection_actor.rs
@@ -354,49 +354,12 @@ async fn push_queue_exhaustion_backpressure() {
 
 use std::sync::{
     Arc,
-    Mutex,
-    OnceLock,
     atomic::{AtomicUsize, Ordering},
 };
 
-use logtest::Logger;
 use serial_test::serial;
 use wireframe::{ConnectionContext, ProtocolHooks};
-
-/// Handle to the global logger with exclusive access.
-struct LoggerHandle {
-    guard: std::sync::MutexGuard<'static, Logger>,
-}
-
-impl LoggerHandle {
-    fn new() -> Self {
-        static LOGGER: OnceLock<Mutex<Logger>> = OnceLock::new();
-
-        let logger = LOGGER.get_or_init(|| Mutex::new(Logger::start()));
-        let guard = logger
-            .lock()
-            .expect("failed to acquire global logger lock; a previous test may still hold it");
-
-        Self { guard }
-    }
-}
-
-impl std::ops::Deref for LoggerHandle {
-    type Target = Logger;
-
-    fn deref(&self) -> &Self::Target { &self.guard }
-}
-
-impl std::ops::DerefMut for LoggerHandle {
-    fn deref_mut(&mut self) -> &mut Self::Target { &mut self.guard }
-}
-
-#[allow(
-    unused_braces,
-    reason = "rustc false positive for single line rstest fixtures"
-)]
-#[fixture]
-fn logger() -> LoggerHandle { LoggerHandle::new() }
+use wireframe_testing::{LoggerHandle, logger};
 
 #[rstest]
 #[tokio::test]
