+++ wireframe_testing/Cargo.toml
@@ -9,3 +9,4 @@ wireframe = { path = ".." }
 bincode = "^2.0"
 bytes = "^1.0"
 rstest = "0.18.2"
+logtest = "2"
