+++ wireframe_testing/src/lib.rs
@@ -14,6 +14,7 @@
 //! ```
 
 pub mod helpers;
+pub mod logging;
 
 pub use helpers::{
     TestSerializer,
@@ -31,3 +32,5 @@ pub use helpers::{
     run_app_with_frames_with_capacity,
     run_with_duplex_server,
 };
+
+pub use logging::{logger, LoggerHandle};
